// planning-idsdp-pps-frontend/src/app/main/modules/rdpp-rtapp-management/models/psc-working-plan.model.ts
export interface PscWorkingPlanModel {
    id: number;
    uuid: string;
    projectConceptMasterId: number;
    projectConceptUuid: string;
    sanctionSchedule: string;
    projectPurpose: string;
    mainActivity: string;
    partWiseExpense: string;
    projectReason: string;
    analysisBackground: string;
    consistencyAnalysis: string;
    projectPrinciple: string;
    analysisType: string;
    rationalityAnalysis: string;
    relatedOtherSubjects: string;
}