// planning-idsdp-pps-frontend/src/app/main/modules/rdpp-rtapp-management/models/dpp-location-wise-cost-breakdown.model.ts
import {IIdUuidHolderRequestBody} from '../../../core/models/request';

export interface IDppLocationWiseCostBreakdown extends IIdUuidHolderRequestBody{

    projectConceptMasterId: number;
    projectConceptMasterUuid: string;
    dppMasterId: number
    rdppMasterId: number
    divisionId: string
    zillaId: string;
    upazilaId: string;
    quantity: string;
    estimatedCost: number;
    comment: string;
}